refactor: vo 패키지 제거 후 vo들을 model 패키지로 이동
