+++ src/main/java/blackjack/controller/BlackjackGame.java
@@ -5,17 +5,17 @@ import blackjack.dto.PlayerBettingProfitOutcome;
 import blackjack.dto.PlayerCardsOutcome;
 import blackjack.dto.PlayerFinalCardsOutcome;
 import blackjack.model.betting.Betting;
+import blackjack.model.betting.BettingMoney;
+import blackjack.model.card.Card;
 import blackjack.model.cardgenerator.CardGenerator;
 import blackjack.model.cardgenerator.RandomCardGenerator;
 import blackjack.model.dealer.Dealer;
 import blackjack.model.player.Player;
+import blackjack.model.player.PlayerName;
 import blackjack.model.player.Players;
 import blackjack.view.InputView;
 import blackjack.view.OutputView;
 import blackjack.view.form.Command;
-import blackjack.vo.BettingMoney;
-import blackjack.vo.Card;
-import blackjack.vo.PlayerName;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
