+++ src/main/java/blackjack/dto/DealerFinalCardsOutcome.java
@@ -1,7 +1,7 @@
 package blackjack.dto;
 
+import blackjack.model.card.Card;
 import blackjack.model.dealer.Dealer;
-import blackjack.vo.Card;
 import java.util.List;
 
 public record DealerFinalCardsOutcome(List<Card> cards, int score) {
