+++ src/main/java/blackjack/dto/PlayerBettingProfitOutcome.java
@@ -3,7 +3,7 @@ package blackjack.dto;
 import blackjack.model.betting.Betting;
 import blackjack.model.dealer.Dealer;
 import blackjack.model.player.Player;
-import blackjack.vo.PlayerName;
+import blackjack.model.player.PlayerName;
 
 public record PlayerBettingProfitOutcome(PlayerName name, int profit) {
     public static PlayerBettingProfitOutcome from(final Player player, final Betting betting, final Dealer dealer) {
