+++ src/main/java/blackjack/dto/PlayerFinalCardsOutcome.java
@@ -1,8 +1,8 @@
 package blackjack.dto;
 
+import blackjack.model.card.Card;
 import blackjack.model.player.Player;
-import blackjack.vo.Card;
-import blackjack.vo.PlayerName;
+import blackjack.model.player.PlayerName;
 import java.util.List;
 
 public record PlayerFinalCardsOutcome(PlayerName name, List<Card> cards, int score) {
