+++ src/main/java/blackjack/model/betting/Betting.java
@@ -1,8 +1,7 @@
 package blackjack.model.betting;
 
 import blackjack.model.player.MatchResult;
-import blackjack.vo.BettingMoney;
-import blackjack.vo.PlayerName;
+import blackjack.model.player.PlayerName;
 import java.util.HashMap;
 import java.util.Map;
