+++ src/main/java/blackjack/model/betting/BettingMoney.java
@@ -1,4 +1,4 @@
-package blackjack.vo;
+package blackjack.model.betting;
 
 public record BettingMoney(int amount) {
     private static final String INVALID_BETTING_MONEY = "베팅 금액은 0원보다 커야 합니다.";
