+++ src/main/java/blackjack/model/card/Card.java
@@ -1,0 +1,4 @@
+package blackjack.model.card;
+
+public record Card(Suit suit, Denomination denomination) {
+}
