+++ src/main/java/blackjack/model/card/Cards.java
@@ -1,6 +1,5 @@
 package blackjack.model.card;
 
-import blackjack.vo.Card;
 import java.util.ArrayList;
 import java.util.List;
