+++ src/main/java/blackjack/model/cardgenerator/CardGenerator.java
@@ -1,6 +1,6 @@
 package blackjack.model.cardgenerator;
 
-import blackjack.vo.Card;
+import blackjack.model.card.Card;
 
 @FunctionalInterface
 public interface CardGenerator {
