+++ src/main/java/blackjack/model/player/PlayerName.java
@@ -1,4 +1,4 @@
-package blackjack.vo;
+package blackjack.model.player;
 
 public record PlayerName(String name) {
     private static final String INVALID_NAME_LENGTH = "플레이어 이름은 한 글자 이상이어야 합니다.";
