+++ src/main/java/blackjack/model/player/Players.java
@@ -1,7 +1,6 @@
 package blackjack.model.player;
 
 import blackjack.model.cardgenerator.CardGenerator;
-import blackjack.vo.PlayerName;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
