+++ src/main/java/blackjack/view/OutputView.java
@@ -4,8 +4,8 @@ import blackjack.dto.DealerFinalCardsOutcome;
 import blackjack.dto.PlayerBettingProfitOutcome;
 import blackjack.dto.PlayerCardsOutcome;
 import blackjack.dto.PlayerFinalCardsOutcome;
-import blackjack.vo.Card;
-import blackjack.vo.PlayerName;
+import blackjack.model.card.Card;
+import blackjack.model.player.PlayerName;
 import java.util.List;
 import java.util.stream.Collectors;
