--- src/main/java/blackjack/vo/Card.java
@@ -1,7 +1,0 @@
-package blackjack.vo;
-
-import blackjack.model.card.Denomination;
-import blackjack.model.card.Suit;
-
-public record Card(Suit suit, Denomination denomination) {
-}
