+++ src/test/java/blackjack/model/card/CardTest.java
@@ -2,7 +2,6 @@ package blackjack.model.card;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import blackjack.vo.Card;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
