+++ src/test/java/blackjack/model/card/CardsTest.java
@@ -2,7 +2,6 @@ package blackjack.model.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import blackjack.vo.Card;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
