+++ src/test/java/blackjack/model/cardgenerator/SequentialCardGenerator.java
@@ -1,6 +1,6 @@
 package blackjack.model.cardgenerator;
 
-import blackjack.vo.Card;
+import blackjack.model.card.Card;
 import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
